fixed logout
git-svn-id: http://localhost:88/svn/uwow4/uwow2/trunk@487 3b753528-4e94-6941-993a-e4fe79dda6b7

+++ src/RealmServer.Core/Gossip/QuestsMenuItem.cs
@@ -4,7 +4,7 @@ namespace Hazzik.Gossip {
 	public class QuestsMenuItem {
 		public QuestsMenuItem(uint menuId, uint icon, string text) {
 			Id = menuId;
-			Icon = (uint)icon;
+			Icon = icon;
 			Text = text;
 		}
